Don't add Python dep when cross-compiling by skipping compiling pyc.
There is only one *.py file currently:
    share/bakefile/presets/wx_presets.py
and it's really not necessary to compile it to a pyc (especially because
it's doubtful anyone is even using these bakefile files).

This means we can avoid having a python dependency added when
cross-compiling, and thus we remove a requirement that the
cross-compiled packages need to be installed with the same Python
interpreter that happened to be used during the compilation.